Relocating classes.
