+++ src/AutomationFoundation.NETFX.App/Hosting/TaskKillRuntimeHostRunAsyncOptions.cs
@@ -1,4 +1,4 @@
-﻿namespace AutomationFoundation
+﻿namespace AutomationFoundation.Hosting
 {
     /// <summary>
     /// Defines the options for an <see cref="TaskKillRuntimeHostRunAsyncStrategy"/>.
