+++ src/AutomationFoundation/Hosting/CtrlCRuntimeHostRunAsyncStrategy.cs
@@ -3,7 +3,7 @@ using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 
-namespace AutomationFoundation
+namespace AutomationFoundation.Hosting
 {
     /// <summary>
     /// Provides a run strategy which runs the host until CTRL+C has been pressed.
