+++ src/AutomationFoundation/Hosting/RuntimeHostRunAsyncOptions.cs
@@ -1,6 +1,6 @@
 ﻿using System.Threading;
 
-namespace AutomationFoundation
+namespace AutomationFoundation.Hosting
 {
     /// <summary>
     /// Defines the options for an <see cref="RuntimeHostRunAsyncStrategy"/>.
